scraper/pcvideoscrapper/clips-api/pyproject.toml
[project]
name = "clips-api"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.25",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
